Refactor module petitions.demarches_simplifiees to petitions.demarche_numerique
