+++ .secrets.baseline
@@ -91,10 +91,6 @@
       "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
     },
     {
-      "path": "detect_secrets.filters.common.is_baseline_file",
-      "filename": ".secrets.baseline"
-    },
-    {
       "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
       "min_level": 2
     },
@@ -188,59 +184,27 @@
         "line_number": 50
       }
     ],
-    "envergo/petitions/demarches_simplifiees/data/fake_dossier.json": [
+    "envergo/petitions/demarche_numerique/data/fake_dossier.json": [
       {
         "type": "Base64 High Entropy String",
-        "filename": "envergo/petitions/demarches_simplifiees/data/fake_dossier.json",
+        "filename": "envergo/petitions/demarche_numerique/data/fake_dossier.json",
         "hashed_secret": "4872d2c6996f10c529ff2a2c8124c0f41c5a22b6",
         "is_verified": false,
         "line_number": 460
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "envergo/petitions/demarches_simplifiees/data/fake_dossier.json",
+        "filename": "envergo/petitions/demarche_numerique/data/fake_dossier.json",
         "hashed_secret": "674db3add05ca81b1c47a0ff413610a8248b5288",
         "is_verified": false,
         "line_number": 467
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "envergo/petitions/demarches_simplifiees/data/fake_dossier.json",
+        "filename": "envergo/petitions/demarche_numerique/data/fake_dossier.json",
         "hashed_secret": "d07bef1bc2c4d23a718bdf8d9af022a5aa5b6e78",
         "is_verified": false,
         "line_number": 491
-      }
-    ],
-    "envergo/petitions/demarches_simplifiees/data/fake_dossier_messages.json": [
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "envergo/petitions/demarches_simplifiees/data/fake_dossier_messages.json",
-        "hashed_secret": "4872d2c6996f10c529ff2a2c8124c0f41c5a22b6",
-        "is_verified": false,
-        "line_number": 51
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "envergo/petitions/demarches_simplifiees/data/fake_dossier_messages.json",
-        "hashed_secret": "674db3add05ca81b1c47a0ff413610a8248b5288",
-        "is_verified": false,
-        "line_number": 58
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "envergo/petitions/demarches_simplifiees/data/fake_dossier_messages.json",
-        "hashed_secret": "d07bef1bc2c4d23a718bdf8d9af022a5aa5b6e78",
-        "is_verified": false,
-        "line_number": 82
-      }
-    ],
-    "envergo/petitions/demarches_simplifiees/models.py": [
-      {
-        "type": "Secret Keyword",
-        "filename": "envergo/petitions/demarches_simplifiees/models.py",
-        "hashed_secret": "5baa61e4c9b93f3f0682250b6cf8331b7ee68fd8",
-        "is_verified": false,
-        "line_number": 129
       }
     ],
     "envergo/users/tests/test_views.py": [
