+++ envergo/petitions/tests/test_views.py
@@ -426,9 +426,7 @@ def test_petition_project_instructor_view_requires_authentication(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_petition_project_instructor_notes_view(
     mock_post, haie_user_44, haie_instructor_44, client, site
 ):
@@ -475,9 +473,7 @@ def test_petition_project_instructor_notes_view(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_instructor_notes_coordinator_sees_both_fields_and_can_edit(
     mock_post, haie_instructor_44, client, site
 ):
@@ -518,9 +514,7 @@ def test_instructor_notes_coordinator_sees_both_fields_and_can_edit(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_instructor_notes_invited_sees_only_public_and_cannot_edit(
     mock_post, haie_user_44, client, site
 ):
@@ -566,9 +560,7 @@ def test_instructor_notes_invited_sees_only_public_and_cannot_edit(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_instructor_notes_invited_empty_notes(mock_post, haie_user_44, client, site):
     """When public notes are empty, invited instructor sees 'Aucune note.'"""
     mock_post.return_value = GET_DOSSIER_FAKE_RESPONSE["data"]
@@ -591,9 +583,7 @@ def test_instructor_notes_invited_empty_notes(mock_post, haie_user_44, client, s
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_instructor_notes_invited_links_are_clickable(
     mock_post, haie_user_44, client, site
 ):
@@ -615,9 +605,7 @@ def test_instructor_notes_invited_links_are_clickable(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_instructor_notes_coordinator_links_are_clickable(
     mock_post, haie_instructor_44, client, site
 ):
@@ -641,9 +629,7 @@ def test_instructor_notes_coordinator_links_are_clickable(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_instructor_notes_coordinator_empty_notes(
     mock_post, haie_instructor_44, client, site
 ):
@@ -666,9 +652,7 @@ def test_instructor_notes_coordinator_empty_notes(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_petition_project_instructor_view_reglementation_pages(
     mock_post,
     haie_instructor_44,
@@ -756,9 +740,7 @@ def test_petition_project_instructor_view_reglementation_pages(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_regulation_view_includes_config_in_context(
     mock_post,
     haie_instructor_44,
@@ -793,9 +775,7 @@ def test_regulation_view_includes_config_in_context(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_petition_project_instructor_display_dossier_ds_info(
     mock_post, haie_instructor_44, client, site
 ):
@@ -823,9 +803,7 @@ def test_petition_project_instructor_display_dossier_ds_info(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_petition_project_instructor_messagerie_ds(
     mock_ds_query_execute, haie_user_44, haie_instructor_44, client, site
 ):
@@ -930,9 +908,7 @@ def _setup_messagerie(haie_instructor_44, client):
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_valid_message_without_file(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -946,9 +922,7 @@ def test_messagerie_valid_message_without_file(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_valid_message_with_file(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -965,9 +939,7 @@ def test_messagerie_valid_message_with_file(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_empty_message_error(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -983,9 +955,7 @@ def test_messagerie_empty_message_error(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_invalid_extension_error(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -1008,9 +978,7 @@ def test_messagerie_invalid_extension_error(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_html_in_filename_is_escaped(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -1026,9 +994,7 @@ def test_messagerie_html_in_filename_is_escaped(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_file_too_large_error(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -1046,9 +1012,7 @@ def test_messagerie_file_too_large_error(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_modal_opens_on_error(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -1064,9 +1028,7 @@ def test_messagerie_modal_opens_on_error(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_messagerie_modal_not_opened_on_success(
     mock_ds_query_execute, haie_instructor_44, client, site
 ):
@@ -1649,9 +1611,7 @@ def test_petition_invited_instructor_cannot_see_send_message_button(
 
 
 @override_settings(DEMARCHE_NUMERIQUE=DEMARCHE_NUMERIQUE_FAKE)
-@patch(
-    "envergo.petitions.demarches_simplifiees.client.DemarchesSimplifieesClient.execute"
-)
+@patch("envergo.petitions.demarche_numerique.client.DemarchesSimplifieesClient.execute")
 def test_petition_invited_instructor_cannot_send_message(
     mock_ds_query_execute, client, haie_instructor_44, haie_user
 ):
